Move constraints to part of parameter space
