+++ emukit/core/optimization/__init__.py
@@ -4,7 +4,6 @@
 
 from .acquisition_optimizer import AcquisitionOptimizerBase  # noqa: F401
 from .context_manager import ContextManager  # noqa: F401
-from .constraints import LinearInequalityConstraint, NonlinearInequalityConstraint
 from .gradient_acquisition_optimizer import GradientAcquisitionOptimizer  # noqa: F401
 from .local_search_acquisition_optimizer import LocalSearchAcquisitionOptimizer  # noqa: F401
 from .multi_source_acquisition_optimizer import MultiSourceAcquisitionOptimizer  # noqa: F401
