+++ emukit/core/optimization/gradient_acquisition_optimizer.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import logging
-from typing import List, Optional, Tuple
+from typing import Tuple
 
 import numpy as np
 
@@ -10,7 +10,6 @@ from .. import ParameterSpace
 from ..acquisition import Acquisition
 from .acquisition_optimizer import AcquisitionOptimizerBase
 from .anchor_points_generator import ConstrainedObjectiveAnchorPointsGenerator, ObjectiveAnchorPointsGenerator
-from .constraints import IConstraint
 from .context_manager import ContextManager
 from .optimizer import OptLbfgs, OptTrustRegionConstrained, apply_optimizer
 
@@ -21,12 +20,11 @@ class GradientAcquisitionOptimizer(AcquisitionOptimizerBase):
     """ Optimizes the acquisition function using a quasi-Newton method (L-BFGS).
     Can be used for continuous acquisition functions.
     """
-    def __init__(self, space: ParameterSpace, constraints: Optional[List[IConstraint]]=None) -> None:
+    def __init__(self, space: ParameterSpace) -> None:
         """
         :param space: The parameter space spanning the search problem.
         """
         super().__init__(space)
-        self.constraints = constraints
 
     def _optimize(self, acquisition: Acquisition, context_manager: ContextManager) -> Tuple[np.ndarray, np.ndarray]:
         """
@@ -70,13 +68,13 @@ class GradientAcquisitionOptimizer(AcquisitionOptimizerBase):
         return x_min, -fx_min
 
     def _get_anchor_points_generator(self, acquisition):
-        if self.constraints is None:
+        if self.space.constraints is None:
             return ObjectiveAnchorPointsGenerator(self.space, acquisition)
         else:
-            return ConstrainedObjectiveAnchorPointsGenerator(self.space, acquisition, self.constraints)
+            return ConstrainedObjectiveAnchorPointsGenerator(self.space, acquisition, self.space.constraints)
 
     def _get_optimizer(self, context_manager):
-        if self.constraints is None:
+        if self.space.constraints is None:
             return OptLbfgs(context_manager.contextfree_space.get_bounds())
         else:
-            return OptTrustRegionConstrained(context_manager.contextfree_space.get_bounds(), self.constraints)
+            return OptTrustRegionConstrained(context_manager.contextfree_space.get_bounds(), self.space.constraints)
