+++ tests/emukit/core/optimization/test_optimizer.py
@@ -4,8 +4,7 @@ import numpy as np
 
 from emukit.core import ParameterSpace, ContinuousParameter
 from emukit.core.optimization import ContextManager
-from emukit.core.optimization.optimizer import OptLbfgs, OptTrustRegionConstrained, apply_optimizer
-from emukit.core.optimization.constraints import LinearInequalityConstraint
+from emukit.core.optimization.optimizer import OptLbfgs, apply_optimizer
 
 
 @pytest.fixture
